Rename files in rule for eslint-plugin-wantedly
