+++ packages/eslint-plugin-wantedly/index.js
@@ -1,17 +1,17 @@
-const GraphQLOperationName = require("./rules/GraphQLOperationName");
-const GraphQLPascalCaseTypeName = require("./rules/GraphQLPascalCaseTypeName");
-const NexusCamelCaseFieldName = require("./rules/NexusCamelCaseFieldName");
-const NexusPascalCaseTypeName = require("./rules/NexusPascalCaseTypeName");
-const NexusUpperCaseEnumMembers = require("./rules/NexusUpperCaseEnumMembers");
+const graphqlOperationName = require("./rules/graphql-operation-name");
+const graphqlPascalCaseTypeName = require("./rules/graphql-pascal-case-type-name");
+const nexusCamelCaseFieldName = require("./rules/nexus-camel-case-field-name");
+const nexusPascalCaseTypeName = require("./rules/nexus-pascal-case-type-name");
 const nexusTypeDescription = require("./rules/nexus-type-description");
+const nexusUpperCaseEnumMembers = require("./rules/nexus-upper-case-enum-members");
 
 module.exports = {
   rules: {
-    [GraphQLOperationName.RULE_NAME]: GraphQLOperationName.RULE,
-    [GraphQLPascalCaseTypeName.RULE_NAME]: GraphQLPascalCaseTypeName.RULE,
-    [NexusCamelCaseFieldName.RULE_NAME]: NexusCamelCaseFieldName.RULE,
-    [NexusPascalCaseTypeName.RULE_NAME]: NexusPascalCaseTypeName.RULE,
-    [NexusUpperCaseEnumMembers.RULE_NAME]: NexusUpperCaseEnumMembers.RULE,
+    [graphqlOperationName.RULE_NAME]: graphqlOperationName.RULE,
+    [graphqlPascalCaseTypeName.RULE_NAME]: graphqlPascalCaseTypeName.RULE,
+    [nexusCamelCaseFieldName.RULE_NAME]: nexusCamelCaseFieldName.RULE,
+    [nexusPascalCaseTypeName.RULE_NAME]: nexusPascalCaseTypeName.RULE,
     [nexusTypeDescription.RULE_NAME]: nexusTypeDescription.RULE,
+    [nexusUpperCaseEnumMembers.RULE_NAME]: nexusUpperCaseEnumMembers.RULE,
   },
 };
