+++ packages/eslint-plugin-wantedly/rules/graphql-operation-name.js
