+++ packages/eslint-plugin-wantedly/rules/nexus-camel-case-field-name.js
@@ -8,7 +8,7 @@ const RULE_NAME = "nexus-camel-case-field-name";
 const WHITELIST_FOR_TYPE_DEFINITION = ["objectType", "interfaceType", "inputObjectType"];
 const FIELD_DEFINITION_METHODS = ["string", "int", "boolean", "id", "float", "field"];
 
-// Represents the default option and schema for graphql-operation-name option
+// Represents the default option and schema for nexus-camel-case-field-name option
 const DEFAULT_OPTION = {
   autofix: false,
 };
